[#2658] add big integer support to statistics
Pool capacities have been promoted in code from uint64_t to uint128_t.

Bigint is modelled as an int128_t in statistics, following the precedent of
signed integer of int64_t from statistics.

- "total-nas" is extended to 128 bits as requested in the issue.
- "total-pds" has the same risk of overflowing so it has been extended
  as well.
- "total-addresses" always fits in 64 bits, but certain code forces a
  128 bit value on it. See Pool::getCapacity(), Subnet::getPoolCapacity(),
  Subnet::sumPoolCapacity(). It could have been truncated to a 64 bit value,
  but that seems like an unnecessary complication.

Because of the disparity in signedness there is some truncation that can
happen when pool capacity values are passed on to statistics. That only happens
for the last half of the value range, so for prefix ranges larger than /1.

+++ src/lib/dhcpsrv/cfg_subnets6.cc
@@ -450,13 +450,11 @@ CfgSubnets6::updateStatistics() {
 
         stats_mgr.setValue(StatsMgr::generateName("subnet", subnet_id,
                                                   "total-nas"),
-                           static_cast<int64_t>
-                           (subnet6->getPoolCapacity(Lease::TYPE_NA)));
+                           subnet6->getPoolCapacity(Lease::TYPE_NA));
 
         stats_mgr.setValue(StatsMgr::generateName("subnet", subnet_id,
                                                   "total-pds"),
-                            static_cast<int64_t>
-                            (subnet6->getPoolCapacity(Lease::TYPE_PD)));
+                           subnet6->getPoolCapacity(Lease::TYPE_PD));
 
         const std::string& name_nas =
             StatsMgr::generateName("subnet", subnet_id, "cumulative-assigned-nas");
